+++ composeApp/src/commonMain/kotlin/com/apimorlabs/reluct/screens/dashboard/overview/DashboardOverviewUI.kt
@@ -56,11 +56,11 @@ import com.apimorlabs.reluct.features.dashboard.states.TodayScreenTimeState
 import com.apimorlabs.reluct.features.dashboard.states.TodayTasksState
 import com.apimorlabs.reluct.screens.dashboard.components.getScreenTimePieChartData
 import com.apimorlabs.reluct.screens.screentime.components.UsagePermissionDialog
-import com.apimorlabs.reluct.screens.util.BottomBarVisibilityHandler
-import com.apimorlabs.reluct.screens.util.GetPermissionsManager
-import com.apimorlabs.reluct.screens.util.PermissionCheckHandler
-import com.apimorlabs.reluct.screens.util.PermissionsManager
-import com.apimorlabs.reluct.screens.util.getSnackbarModifier
+import com.apimorlabs.reluct.util.BottomBarVisibilityHandler
+import com.apimorlabs.reluct.util.GetPermissionsManager
+import com.apimorlabs.reluct.util.PermissionCheckHandler
+import com.apimorlabs.reluct.util.PermissionsManager
+import com.apimorlabs.reluct.util.getSnackbarModifier
 import kotlinx.coroutines.launch
 import org.jetbrains.compose.resources.painterResource
 import org.jetbrains.compose.resources.stringResource
