+++ composeApp/src/desktopMain/kotlin/com/apimorlabs/reluct/screens/util/PermissionsManager.desktop.kt
@@ -3,6 +3,7 @@ package com.apimorlabs.reluct.screens.util
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.LaunchedEffect
 import androidx.compose.runtime.remember
+import com.apimorlabs.reluct.util.PermissionsManager
 
 class PermissionsManagerDesktop : PermissionsManager {
     override fun checkUsageAccessPermission(): Boolean = true
